Support customized HostMapping implementation
