+++ src/main/java/org/tikv/common/TiConfiguration.java
@@ -255,6 +255,7 @@ public class TiConfiguration implements Serializable {
   private int metricsPort = getInt(TIKV_METRICS_PORT);
 
   private final String networkMappingName = get(TIKV_NETWORK_MAPPING_NAME);
+  private HostMapping hostMapping = null;
 
   public enum KVMode {
     TXN,
@@ -531,5 +532,13 @@ public class TiConfiguration implements Serializable {
 
   public String getNetworkMappingName() {
     return this.networkMappingName;
+  }
+
+  public HostMapping getHostMapping() {
+    return hostMapping;
+  }
+
+  public void setHostMapping(HostMapping mapping) {
+    this.hostMapping = mapping;
   }
 }
